Updates heap dump logic to inspect environment variable.
Per a discussion in-person with Steven during review, the argument-less
`dumpHeap` method now conditions on the presence of `$MESOS_SANDBOX`.
If present, it'll attempt to dump heap there.  If not, it will try to
get the system temporary directory.  If that fails, it will try `/tmp`.

The test for inspecting the Mesos environment variable isn't
parallel-safe.  Perhaps that's ok; it hardly seems worth it to convert
this all to use DI or at least require instances.